:sparkles:  printing to table enabled
:sparkles:
as long as your in the folder with these files, you can run the following in the matlab command window:
b = system();
runpf(b);
and you will get an excel file, testitesti4.xlsx, which contains what you want.
take that file, save it somewhere else, then delete it from the folder and repeat as needed.